Separated constants into separate files
