+++ frontend/lib/constants/intros.ts
@@ -1,0 +1,57 @@
+import type { Skill } from "../types";
+
+export const intros: Skill[] = [
+  {
+    title: "a full-stack Software Developer",
+    icon: "fluent-emoji:man-technologist-light",
+  },
+  {
+    title: "a full-stack Web Developer",
+    icon: "fluent-color:code-20",
+  },
+  {
+    title: "a JavaScript Developer",
+    icon: "catppuccin:javascript",
+  },
+  { title: "a Python Developer", icon: "catppuccin:python" },
+  { title: "a PHP Developer", icon: "catppuccin:php" },
+  { title: "a React Developer", icon: "catppuccin:javascript-react" },
+  { title: "a Vue Developer", icon: "catppuccin:vue" },
+  {
+    title: "an Angular Developer",
+    icon: "catppuccin:angular",
+  },
+  { title: "a Django Developer", icon: "catppuccin:django" },
+  {
+    title: "a Laravel Developer",
+    icon: "catppuccin:laravel",
+  },
+  {
+    title: "a WebGL and Three.js Developer",
+    icon: "skill-icons:threejs-dark",
+  },
+  { title: "a VR/AR Developer", icon: "fluent-emoji:goggles" },
+  {
+    title: "a teacher of Web Programming",
+    icon: "fluent-emoji:man-teacher-light",
+  },
+  { title: "a web designer", icon: "fluent-emoji:spider-web" },
+  { title: "a graphic designer", icon: "fluent-color:design-ideas-20" },
+  { title: "a UI/UX designer", icon: "fluent-color:phone-laptop-20" },
+  { title: "an illustrator", icon: "fluent-emoji:man-artist-light" },
+  { title: "a cartoonist", icon: "fluent-emoji:pencil" },
+  {
+    title: "a college graduate in Software Development",
+    icon: "fluent-emoji:student-light",
+  },
+  {
+    title: "living in Ontario, Canada",
+    icon: "cif:ca",
+  },
+  {
+    title: "a pretty cool guy (I think)",
+    icon: "fluent-emoji:smiling-face-with-sunglasses",
+  },
+];
+
+export default intros;
