+++ frontend/lib/constants/lottieConstants.ts
@@ -1,0 +1,41 @@
+import type { LottieConstants } from "../types";
+
+export const lottieConstants: LottieConstants = {
+  CLOUDS_LIGHT_LOTTIE_PATH: "/lottie/bg/clouds_light.json",
+  CLOUDS_DARK_LOTTIE_PATH: "/lottie/bg/clouds_dark.json",
+  AVATAR_WAVE_LOTTIE_PATH: "/lottie/avatars/timrlai_avatar_wave.json",
+  AVATAR_SUMMARY_LOTTIE_PATH: "/lottie/avatars/timrlai_avatar_summary.json",
+  AVATAR_SKILLS_DESK_LOTTIE_PATH:
+    "/lottie/avatars/timrlai_avatar_skills_desk.json",
+  AVATAR_SKILLS_LEGS_LOTTIE_PATH:
+    "/lottie/avatars/timrlai_avatar_skills_legs.json",
+  AVATAR_SKILLS_SOFT_LOTTIE_PATH:
+    "/lottie/avatars/timrlai_avatar_skills_soft.json",
+  AVATAR_VR_LOTTIE_PATH: "/lottie/avatars/timrlai_avatar_vr.json",
+  AVATAR_TEACHER_DESK_LOTTIE_PATH:
+    "/lottie/avatars/timrlai_avatar_teacher_desk.json",
+  AVATAR_TEACHER_LEGS_LOTTIE_PATH:
+    "/lottie/avatars/timrlai_avatar_teacher_legs.json",
+  AVATAR_DESIGNER_DESK_LOTTIE_PATH:
+    "/lottie/avatars/timrlai_avatar_designer_desk.json",
+  AVATAR_DESIGNER_NO_DESK_LOTTIE_PATH:
+    "/lottie/avatars/timrlai_avatar_designer_no_desk.json",
+  AVATAR_ILLUSTRATOR_LOTTIE_PATH:
+    "/lottie/avatars/timrlai_avatar_illustrator.json",
+  AVATAR_EDUCATION_LOTTIE_PATH: "/lottie/avatars/timrlai_avatar_education.json",
+  AVATAR_VOLUNTEER_LOTTIE_PATH: "/lottie/avatars/timrlai_avatar_volunteer.json",
+  BAT_LOTTIE_PATH: "/lottie/bats/timrlai_bat.json",
+  BAT_SUMMARY_LOTTIE_PATH: "/lottie/bats/timrlai_bat_summary.json",
+  BAT_SKILLS_LOTTIE_PATH: "/lottie/bats/timrlai_bat_skills.json",
+  BAT_SKILLS_LAPTOP_LOTTIE_PATH: "/lottie/bats/timrlai_bat_skills_laptop.json",
+  BAT_SKILLS_SOFT_LOTTIE_PATH: "/lottie/bats/timrlai_bat_skills_soft.json",
+  NOT_FOUND_LOTTIE_FOLDER: "/lottie/404/",
+  NOT_FOUND_LOTTIES: [
+    "404_cat_clouds.json",
+    "404_cat_magnifying_glass.json",
+    "404_cat_papers.json",
+    "404_cat_yarn.json",
+  ],
+};
+
+export default lottieConstants;
