+++ frontend/lib/constants/socialButtons.ts
@@ -1,0 +1,66 @@
+import type { SocialButton } from "../types";
+
+export const socialButtons: SocialButton[] = [
+  {
+    tooltip: "Twitch",
+    url: "https://twitch.tv/timrlai",
+    icon: "cib:twitch",
+    main: false,
+  },
+  {
+    tooltip: "YouTube",
+    url: "https://www.youtube.com/@timrlai",
+    icon: "mingcute:youtube-fill",
+    main: false,
+  },
+  {
+    tooltip: "Vimeo",
+    url: "https://vimeo.com/timrlai",
+    icon: "entypo-social:vimeo-with-circle",
+    main: false,
+  },
+  {
+    tooltip: "Artstation",
+    url: "https://artstation.com/timrlai",
+    icon: "cib:artstation",
+    main: false,
+  },
+  {
+    tooltip: "Dribbble",
+    url: "https://dribbble.com/timrlai",
+    icon: "mingcute:dribbble-fill",
+    main: false,
+  },
+  {
+    tooltip: "Instagram",
+    url: "https://instagram.com/timrlai",
+    icon: "mingcute:instagram-line",
+    main: false,
+  },
+  {
+    tooltip: "Threads",
+    url: "https://threads.com/timrlai",
+    icon: "mingcute:threads-fill",
+    main: false,
+  },
+  {
+    tooltip: "Twitter (No, not X)",
+    url: "https://twitter.com/timrlai",
+    icon: "mingcute:twitter-fill",
+    main: false,
+  },
+  {
+    tooltip: "GitHub",
+    url: "https://github.com/timrlai",
+    icon: "mingcute:github-fill",
+    main: true,
+  },
+  {
+    tooltip: "LinkedIn",
+    url: "https://www.linkedin.com/in/timrlai/",
+    icon: "mingcute:linkedin-fill",
+    main: true,
+  },
+];
+
+export default socialButtons;
