+++ frontend/lib/constants/subItems.ts
@@ -1,0 +1,67 @@
+import type { SubItem } from "../types";
+
+export const meetSubItems: SubItem[] = [
+  {
+    label: "Hi!",
+    to: "/",
+    hash: "hi",
+    icon: "mingcute:hand-fill",
+  },
+  {
+    label: "Summary",
+    to: "/",
+    hash: "summary",
+    icon: "mingcute:list-check-3-fill",
+  },
+  {
+    label: "Skills",
+    to: "/",
+    hash: "skills",
+    icon: "mingcute:code-fill",
+  },
+  {
+    label: "Experience",
+    to: "/",
+    hash: "experience",
+    icon: "mingcute:document-2-fill",
+  },
+];
+
+export const experienceSubItems: SubItem[] = [
+  {
+    label: "Developer",
+    to: "/experience",
+    hash: "developer",
+    icon: "mingcute:code-fill",
+  },
+  {
+    label: "Teacher",
+    to: "/experience",
+    hash: "teacher",
+    icon: "mingcute:school-fill",
+  },
+  {
+    label: "Designer",
+    to: "/experience",
+    hash: "designer",
+    icon: "mingcute:palette-fill",
+  },
+  {
+    label: "Illustrator",
+    to: "/experience",
+    hash: "illustrator",
+    icon: "mingcute:paint-brush-fill",
+  },
+  {
+    label: "Education",
+    to: "/experience",
+    hash: "education",
+    icon: "mingcute:mortarboard-fill",
+  },
+  {
+    label: "Volunteer",
+    to: "/experience",
+    hash: "volunteer",
+    icon: "mingcute:hand-heart-fill",
+  },
+];
