+++ frontend/lib/constants/ui.ts
@@ -1,0 +1,26 @@
+export const buttonColors = {
+  neutral: "btn-neutral",
+  primary: "btn-primary",
+  secondary: "btn-secondary",
+  accent: "btn-accent",
+  info: "btn-info",
+  success: "btn-success",
+  warning: "btn-warning",
+  error: "btn-error",
+};
+export const tooltipColors = {
+  neutral: "tooltip-neutral",
+  primary: "tooltip-primary",
+  secondary: "tooltip-secondary",
+  accent: "tooltip-accent",
+  info: "tooltip-info",
+  success: "tooltip-success",
+  warning: "tooltip-warning",
+  error: "tooltip-error",
+};
+export const tooltipPositions = {
+  left: "tooltip-left",
+  right: "tooltip-right",
+  top: "tooltip-top",
+  bottom: "tooltip-bottom",
+};
